+++ packages/editor-common/web/src/consts.ts
@@ -163,7 +163,7 @@ export const INSERT_PLUGIN_BUTTONS = Object.freeze({
   COLLAPSIBLE_LIST: 'CollapsibleList_InsertButton',
   ADSENSE: 'AdSensePlugin_InsertButton',
   AUDIO: 'AudioPlugin_InsertButton',
-  SPOTIFY: 'SpotifyPlugin_InsertButton',
+  SPOTIFY: 'Spotify_InsertButton',
 });
 
 export const BUTTON_TYPES = Object.freeze({