+++ packages/plugin-audio/web/src/modals.ts
@@ -1,0 +1,11 @@
+import AudioModal from './toolbar/modals/AudioModal';
+
+const Modals = {
+  AUDIO_MODAL: 'audio',
+};
+
+const ModalsMap = {
+  [Modals.AUDIO_MODAL]: AudioModal,
+};
+
+export { Modals, ModalsMap };