+++ packages/ui-components/web/src/Icons/index.ts
@@ -17,6 +17,8 @@ export { default as ContextMenuIcon } from './ContextMenuIcon';
 export { default as PlaybackIcon } from './PlaybackIcon';
 export { default as ReplaceIcon } from './ReplaceIcon';
 export { default as DeleteIcon } from './DeleteIcon';
+export { default as SoundCloudIcon } from './SoundCloudIcon';
+export { default as SpotifyIcon } from './SpotifyIcon';
 
 //need to import from plugin-commons
 // export { default as AlignCenterIcon } from './AlignCenterIcon';